Add parsing.py
Initialisation of script.

Script uses Stanza to access the Alpino parsing treebank.

Includes count_phrase_incidence() function to count the number of times a specific POS tag is used
and to generate a list with the words that triggered this tag.

Includes apply_incidence_count() function to apply previous function according to a predefined set
of POS tags and to return a list with tuples with the indices in design.