Add get_synt_feat() function
Similar to the function in this version of parsing.py, but this time working with the output of Alpino
and with the dependency tags instead of the POS tags.

Extracts the number of vp, np and pp in each text and returns a tuple with a list of all the phrase
instances (not only these three), then the count of vp, np and pp.

To be corrected: Does not recognise vps since they are not tagged as such.